Moving ATTRIBUTE into src/types
